+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated how the exception handler responds as JSON.
 - All Bones services are now being saved in the container as the fully qualified class name to utilize 
 dependency injection.
+- Updated how the `App::start()` method loads routes by adding a private `loadRoutes` method.
 
 ### Removed